requirements.txt
numpy~=1.23.2
Pillow~=7.0.0
matplotlib~=3.5.3
scikit-learn~=1.1.2
torch~=1.12.1
PyYAML~=5.3.1
requests~=2.26.0
scipy~=1.9.1
python-multipart==0.0.5
pydantic==1.10.2
imutils==0.5.4
onnxruntime==1.12.1
onnx==1.12.0
coremltools>=4.1
onnxruntime
protobuf<4.21.3
onnxruntime-gpu
onnx>=1.9.0
onnx-simplifier>=0.3.6
fastapi==0.82.0
tabulate==0.8.10
tqdm==4.64.1
plyer==2.0.0
service==0.6.0
uvicorn==0.18.3
starlette==0.19.1
ipython==8.5.0
imagezmq==1.1.1
scikit-image==0.19.3